chore(reorganize): organize project files
