+++ .gitignore
@@ -3,7 +3,7 @@ OpenSprinkler
 wtopts.txt
 *.dat
 *.sh
-testmode.h
+include/testmode.h
 build-1284/*
 .vscode
 .pio
