+++ src/TimeLib.cpp
@@ -33,8 +33,8 @@
 #include <WProgram.h>
 #endif
 
-#include "TimeLib.h"
-#include "types.h"
+#include "../include/TimeLib.h"
+#include "../include/types.h"
 
 static tmElements_t tm;  // a cache of time elements
 static time_os_t cacheTime;  // the time the cache was updated
