+++ src/program.cpp
@@ -22,8 +22,8 @@
  */
 
 #include <limits.h>
-#include "program.h"
-#include "main.h"
+#include "../include/program.h"
+#include "../include/main.h"
 
 #if !defined(SECS_PER_DAY)
 #define SECS_PER_MIN  (60UL)
