+++ src/smtp.c
@@ -62,7 +62,7 @@
  */
 #define SMTP_INTERNAL_DEFINE
 
-#include "smtp.h"
+#include "../include/smtp.h"
 
 /*
  * The SMTP_TEST converts some library routines into special test seams which
